Fixed the log in and combined the 2 app.py and restructured everything
