--- init_db.py
@@ -1,6 +1,0 @@
-from app import app, db
-
-with app.app_context():
-    # Create all tables
-    db.create_all()
-    print("Database tables created successfully!")
