+++ src/foodscanner/static/Sample.css
