+++ src/foodscanner/static/auth.css
