+++ src/foodscanner/static/dashboard.css
@@ -121,26 +121,109 @@ body {
 
 .countdown {
     display: flex;
+    align-items: center;
     justify-content: center;
-    gap: 2rem;
-    margin-top: 1rem;
+    gap: 1rem;
+    margin: 1rem 0;
 }
 
 .countdown-item {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+    text-align: center;
+    min-width: 80px;
 }
 
-.countdown-item .number {
-    font-size: 2.5rem;
+.countdown-separator {
+    font-size: 2rem;
     font-weight: bold;
     color: var(--primary-color);
 }
 
-.countdown-item .label {
+.number {
+    font-size: 2rem;
+    font-weight: bold;
+    color: var(--primary-color);
+    display: block;
+}
+
+.label {
+    font-size: 0.8rem;
+    color: var(--text-color);
+    opacity: 0.8;
+}
+
+.due-date-text {
+    text-align: center;
+    margin-top: 0.5rem;
     font-size: 0.9rem;
-    color: #666;
+    color: var(--text-color);
+}
+
+.water-controls {
+    display: flex;
+    gap: 0.5rem;
+    margin-top: 1rem;
+}
+
+.btn-water {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.5rem;
+    padding: 0.5rem 1rem;
+    border: none;
+    border-radius: 5px;
+    background-color: var(--primary-color);
+    color: white;
+    font-size: 0.9rem;
+    cursor: pointer;
+    transition: all 0.3s ease;
+}
+
+.btn-water:hover {
+    background-color: var(--primary-dark);
+    transform: translateY(-2px);
+}
+
+.btn-water:active {
+    transform: translateY(0);
+}
+
+.btn-water i {
+    font-size: 0.8rem;
+}
+
+/* Add hover effect for the progress bar */
+.progress-bar {
+    position: relative;
+    overflow: hidden;
+}
+
+.progress-bar:hover .progress {
+    opacity: 0.8;
+}
+
+/* Add animation for progress updates */
+.progress {
+    transition: width 0.3s ease, opacity 0.3s ease;
+}
+
+/* Add styles for the stat value */
+.stat-value {
+    display: flex;
+    align-items: baseline;
+    gap: 0.5rem;
+    margin: 0.5rem 0;
+}
+
+.stat-value span {
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: var(--primary-color);
+}
+
+.stat-value .stat-label {
+    font-size: 1rem;
+    color: var(--text-color);
+    opacity: 0.8;
 }
 
 .trimester-info {
