+++ src/foodscanner/static/log_meal.css
