+++ src/foodscanner/static/profile.js
