+++ src/foodscanner/static/recommendations.js
