+++ src/foodscanner/static/script.js
