+++ src/foodscanner/templates/Sample.html
