+++ src/foodscanner/templates/auth.html
@@ -33,11 +33,11 @@
                 <form id="loginForm">
                     <div class="form-group">
                         <label for="username">Username</label>
-                        <input type="text" id="username" required>
+                        <input type="text" id="username" name="username" required>
                     </div>
                     <div class="form-group">
                         <label for="password">Password</label>
-                        <input type="password" id="password" required>
+                        <input type="password" id="password" name="password" required>
                     </div>
                     <button type="submit" class="btn-primary">Login</button>
                 </form>
