+++ src/foodscanner/templates/recommendations.html
