--- templates/log_meal.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>NutriLogic - Log Your Meal</title>
-    <link rel="icon" type="image/png" sizes="32x32" href="/static/favicon/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="/static/favicon/favicon-16x16.png">
-    <link rel="apple-touch-icon" sizes="180x180" href="/static/favicon/apple-touch-icon.png">
-    <link rel="manifest" href="/static/favicon/site.webmanifest">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css">
-    <link rel="stylesheet" href="static/log_meal.css">
-</head>
-<body>
-    <!-- Navbar -->
-    <nav class="navbar">
-        <div class="container">
-            <div class="logo">
-                <i class="fa-solid fa-leaf"></i>
-                <span class="logo-text"><span class="highlight">Nutri</span>Logic</span>
-            </div>
-            <ul class="nav-links">
-                <li><a href="/dashboard">Dashboard</a></li>
-                <li><a href="/log-meal" class="active">Log Meal</a></li>
-                <li><a href="/chat">Chat</a></li>
-                <li><a href="/meal-plans">Meal Plans</a></li>
-                <li><a href="/resources">Resources</a></li>
-                <li><a href="/logout" class="btn-logout">Logout</a></li>
-            </ul>
-            <div class="menu-toggle">
-                <i class="fa-solid fa-bars"></i>
-            </div>
-        </div>
-    </nav>
-
-    <!-- Main Content -->
-    <div class="container">
-        <div class="meal-log-container">
-            <h1>Log Your Meal</h1>
-            <form id="upload-form" action="/upload" method="POST" enctype="multipart/form-data">
-                <!-- Image Upload Section -->
-                <div class="upload-section">
-                    <h2>Upload an image of your meal</h2>
-                    <div id="upload-box" class="upload-box">
-                        <input type="file" id="meal-upload" name="mealImage" class="hidden" accept="image/*">
-                        <div class="upload-content">
-                            <i class="fa-solid fa-cloud-arrow-up"></i>
-                            <p>Click to upload or drag and drop</p>
-                        </div>
-                        <img id="preview-image" class="preview-image hidden" />
-                    </div>
-                </div>
-
-                <!-- Meal Description Section -->
-                <div class="description-section">
-                    <h2>Describe your meal</h2>
-                    <textarea id="meal-description" name="mealText" placeholder="Write about your meal..."></textarea>
-                </div>
-
-                <!-- Submit Button -->
-                <div class="submit-section">
-                    <button id="submit-button" type="submit" class="btn-submit" disabled>
-                        <i class="fa-solid fa-utensils"></i> Get Details
-                    </button>
-                </div>
-            </form>
-        </div>
-    </div>
-
-    <script src="static/log_meal.js"></script>
-</body>
-</html>
